Fix logging and config
Beofre, logging was configurable through a logging config. From now on,
logging is fixed in the code and a certain way of logging within
supercell application is enforced.

In a request handler you should only have to call `self.logger` to
access a logger, i.e. a `self.logger.info()` would emit an INFO log
message. Along with the message and the exact request handler class
name, each request is given an almost unique ID that should make
debugging easier.

If for what ever reason you are not happy with the decisions made,
Simply override the `Service.initialize_logging` method to do something
else.

Refs #2