Allow easier Application creation
With this change it is now possible to create a
`supercell.api.Application`. Before it was not possible to add the
handlers directly in the c'tor, now you may create one with a mocked
environment and/or config:

    app = Application(mocked_environment, mocked_config, handlers)

where everything after the `mocked_config` is the standard
`tornado.web.Application` instantiation.