Supercell stats using greplin scales
Two new decorators allow for collecting stats for your request handlers:
s.latency measures execution latency of a certain method in your request
handler and s.metered tracks 1/5/15 minutes averages on the method
calls.

In order to access the stats, */_system/stats/* returns the overall
stats. You may add more request paths to the stats handler in order to
get only those parts of requests, you are interested in.

So if you define a request handler in the */api/v1/echo* route and use
the stats decorators, they will be recorded with the complete route, so
accessing only stats for this handler is done with
*/_system/stats/api/v1/echo*.

Closes #3