Allow rendering HTML using a simple provider
This patch introduces a simple HTML rendering mechanism for supercell. A
custom `TornadoTemplateProvider` allows for rendering using tornado's
built in template language. In order to determine the template to use,
the `RequestHandler` has to implement the `get_template()` method that
takes the final `model` as argument. The returned string must be a
filename relative to the `template_path` tornado setting.